+++ Application/Common/Model/DepartmentModel.class.php
@@ -5,11 +5,11 @@
  * @Email:369709991@qq.com
  * @Date:   2017-07-11 12:15:19
  * @Last Modified by:   vition
- * @Last Modified time: 2017-07-13 18:52:40
+ * @Last Modified time: 2017-07-24 12:56:44
  */
 namespace Common\Model;
-use Think\Model;
-class DepartmentModel extends Model{
+use Common\Model\AmongModel;
+class DepartmentModel extends AmongModel{
 	protected $trueTableName = 'oa_department'; 
 	protected $fields = array('department_id', 'department_name','department_leader');
 
@@ -21,7 +21,9 @@ class DepartmentModel extends Model{
 	 * @return [type]        [上述两个参数都为空的时候，默认查询所有；$start存在而$limit为空的时候，查询条数；两个参数都存在则查询指定起始和限制条数]
 	 */
 	function search_department($department_leader=0,$start=0,$limit=0,$condition=""){
-
+		if(!$this->has_auth("select")){
+			return false;
+		}
 		if(is_array($condition)){
 			$this->where($condition);
 		}
@@ -42,6 +44,9 @@ class DepartmentModel extends Model{
 	 * @return [type]                  [当第二个参数不为空的时候第一个参数失效]
 	 */
 	function find_department($department_id,$department_name="",$department_leader=0){
+		if(!$this->has_auth("select")){
+			return false;
+		}
 		if($department_name!=""){
 			$condition['department_name']=$department_name;
 		}else if($department_leader>0){
@@ -58,6 +63,9 @@ class DepartmentModel extends Model{
 	 * @return boolean                    [description]
 	 */
 	function is_department($department_name,$department_leader=0){
+		if(!$this->has_auth("select")){
+			return false;
+		}
 		return $this->where(array("department_name"=>$department_name,"department_leader"=>$department_leader))->find();
 	}
 	/**
@@ -67,6 +75,9 @@ class DepartmentModel extends Model{
 	 * @param [type] $department_name [是否为管理层，默认非]
 	 */
 	function set_department($department_id,$department_name,$department_leader=0){
+		if(!$this->has_auth("update")){
+			return false;
+		}
 		$reaultArray=$this->is_department($department_name,$department_leader);
 		if(empty($reaultArray)){
 			return $this->where(array("department_id"=>$department_id))->save(array("department_name"=>$department_name,"department_leader"=>$department_leader));
@@ -80,6 +91,9 @@ class DepartmentModel extends Model{
 	 * @param [type] $department_name 新增的部门名
 	 */
 	function add_department($department_name,$department_leader=0){
+		if(!$this->has_auth("insert")){
+			return false;
+		}
 		$reaultArray=$this->is_department($department_name,$department_leader);
 		
 		if(empty($reaultArray)){
@@ -95,6 +109,9 @@ class DepartmentModel extends Model{
 	 * @return [type]             
 	 */
 	function del_department($department_id){
+		if(!$this->has_auth("delete")){
+			return false;
+		}
 		return $this->where(array("department_id"=>$department_id))->delete();
 	}
