+++ Application/Common/Model/PlaceModel.class.php
@@ -4,12 +4,12 @@
  * @Email:369709991@qq.com
  * @Date:   2017-07-11 15:01:41
  * @Last Modified by:   vition
- * @Last Modified time: 2017-07-21 10:52:50
+ * @Last Modified time: 2017-07-24 12:57:15
  */
 namespace Common\Model;
-use Think\Model;
+use Common\Model\AmongModel;
 
-class PlaceModel extends Model{
+class PlaceModel extends AmongModel{
 	protected $trueTableName = 'oa_place'; 
 	protected $fields = array('place_id', 'place_name','place_department','place_group','place_manager','place_extent','place_role');
 
@@ -22,6 +22,9 @@ class PlaceModel extends Model{
 	 * @return [type]        [上述两个参数都为空的时候，默认查询所有；$start存在而$limit为空的时候，查询条数；两个参数都存在则查询指定起始和限制条数]
 	 */
 	function search_place($place_department,$place_group=0,$start="",$limit=""){
+		if(!$this->has_auth("select")){
+			return false;
+		}
 		$place=$this->where(array("place_department"=>$place_department,"place_group"=>$place_group));
 		if($start=="" && $limit==""){
 			return $placeData=$place->select();
@@ -38,6 +41,9 @@ class PlaceModel extends Model{
 	 * @return [type]           []
 	 */
 	function find_place($place_id){
+		if(!$this->has_auth("select")){
+			return false;
+		}
 		return $this->where(array("place_id"=>$place_id))->find();
 	}
 
@@ -50,6 +56,9 @@ class PlaceModel extends Model{
 	 * @return boolean                   [description]
 	 */
 	function is_place($place_department,$place_name,$place_manager,$place_group=0){
+		if(!$this->has_auth("select")){
+			return false;
+		}
 		return $this->where(array("place_department"=>$place_department,"place_name"=>$place_name,"place_group"=>$place_group,"place_manager"=>$place_manager))->find();
 	}
 
@@ -63,6 +72,9 @@ class PlaceModel extends Model{
 	 * @param integer $place_group      [职位分组]
 	 */
 	function set_place($place_id,$place_department,$place_name,$place_manager,$place_group=0){
+		if(!$this->has_auth("update")){
+			return false;
+		}
 		$isResult=$this->is_place($place_department,$place_name,$place_manager,$place_group);
 		if($isResult==""){
 			return $this->where(array("place_id"=>$place_id))->save(array("place_name"=>$place_name,"place_department"=>$place_department,"place_group"=>$place_group,"place_manager"=>$place_manager));
@@ -80,6 +92,9 @@ class PlaceModel extends Model{
 	 * @param integer $place_group      [职位分组]
 	 */
 	function add_place($place_department,$place_name,$place_manager,$place_group=0){
+		if(!$this->has_auth("insert")){
+			return false;
+		}
 		if($this->is_place($place_department,$place_name,$place_manager,$place_group)==""){
 			return $this->add(array("place_name"=>$place_name,"place_department"=>$place_department,"place_group"=>$place_group,"place_manager"=>$place_manager));
 		}else{
@@ -93,6 +108,9 @@ class PlaceModel extends Model{
 	 * @return [type]           [description]
 	 */
 	function del_place($place_id){
+		if(!$this->has_auth("delete")){
+			return false;
+		}
 		return $this->where(array("place_id"=>$place_id))->delete();
 	}
 	/**
@@ -100,6 +118,9 @@ class PlaceModel extends Model{
 	 * @return [type] [description]
 	 */
 	function get_leader(){
+		if(!$this->has_auth("select")){
+			return false;
+		}
 		return $this->join("left join oa_department d on d.department_id=place_department")->where("d.department_leader>0 and place_manager>0")->select();
 	}
 
@@ -109,6 +130,9 @@ class PlaceModel extends Model{
 	 * @param [type] $eplace_extent [需要新增的extent]
 	 */
 	function add_extent($place_id,$eplace_extent){
+		if(!$this->has_auth("update")){
+			return false;
+		}
 		$map['place_extent'] = array("exp","concat ('{$eplace_extent}',place_extent)");
 		$this->where(array("place_id"=>$place_id))->save($map);
 	}
@@ -120,6 +144,9 @@ class PlaceModel extends Model{
 	 * @return [type]               [none]
 	 */
 	function reduce_extent($place_id,$place_extent){
+		if(!$this->has_auth("update")){
+			return false;
+		}
 		$this->where(array("place_id"=>$place_id))->save(array("place_extent"=>$place_extent));
 	}
 
@@ -129,6 +156,9 @@ class PlaceModel extends Model{
 	 * @param [type] $place_role [修改的role id]
 	 */
 	function set_place_role($place_id,$place_role){
+		if(!$this->has_auth("update")){
+			return false;
+		}
 		$this->where(array("place_id"=>$place_id))->save(array("place_role"=>$place_role));
 	}
 }
