+++ Application/Common/Model/RoleModel.class.php
@@ -4,15 +4,15 @@
  * @Email:369709991@qq.com
  * @Date:   2017-07-11 15:37:56
  * @Last Modified by:   vition
- * @Last Modified time: 2017-07-21 09:59:31
+ * @Last Modified time: 2017-07-24 12:57:28
  */
 namespace Common\Model;
-use Think\Model;
+use Common\Model\AmongModel;
 
 /**
  * 角色管理模型
  */
-class RoleModel extends Model{
+class RoleModel extends AmongModel{
 	protected $trueTableName = 'oa_role'; 
 	protected $fields = array('role_id', 'role_name','role_upper');
 
@@ -25,6 +25,9 @@ class RoleModel extends Model{
 	 * @return [type]        [上述两个参数都为空的时候，默认查询所有；$start存在而$limit为空的时候，查询条数；两个参数都存在则查询指定起始和限制条数]
 	 */
 	function search_role($role_upper=0,$all_upper=false,$start="",$limit=""){
+		if(!$this->has_auth("select")){
+			return false;
+		}
 		if($all_upper==true){
 			$role=$this;
 		}else{
@@ -46,6 +49,9 @@ class RoleModel extends Model{
 	 * @return [type]           []
 	 */
 	function find_role($role_id){
+		if(!$this->has_auth("select")){
+			return false;
+		}
 		return $this->where(array("role_id"=>$role_id))->find();
 	}
 
@@ -56,6 +62,9 @@ class RoleModel extends Model{
 	 * @return boolean             []
 	 */
 	function is_role($role_name,$role_upper=0){
+		if(!$this->has_auth("select")){
+			return false;
+		}
 		return $this->where(array("role_name"=>$role_name,"role_upper"=>$role_upper))->find();
 	}
 
@@ -66,6 +75,9 @@ class RoleModel extends Model{
 	 * @param integer $role_upper [角色分组]
 	 */
 	function set_role($role_id,$role_name,$role_upper=0){
+		if(!$this->has_auth("update")){
+			return false;
+		}
 		$resultData=$this->is_role($role_name,$role_upper);
 		if($resultData==""){
 			return $this->where(array('role_id' =>$role_id ))->save(array("role_name"=>$role_name,"role_upper"=>$role_upper));
@@ -80,6 +92,9 @@ class RoleModel extends Model{
 	 * @param integer $role_upper [角色分组id]
 	 */
 	function add_role($role_name,$role_upper=0){
+		if(!$this->has_auth("insert")){
+			return false;
+		}
 		if($this->is_role($role_name,$role_upper)==""){
 			return $this->add(array("role_name"=>$role_name,"role_upper"=>$role_upper));
 		}else{
@@ -93,6 +108,9 @@ class RoleModel extends Model{
 	 * @return [type]           [description]
 	 */
 	function del_role($role_id){
+		if(!$this->has_auth("delete")){
+			return false;
+		}
 		return $this->where(array("role_id"=>$role_id))->delete();
 	}
 }
